Expanded on FileOpen/Browser/DragAction UI
Reinstated and revised the content about the drag-and-drop open modes.
Pointed out that these modes only apply to drag-and-drop.  Added info
about how to work the mode-switches (mid-drag and drop methods are a
nice design feature in the software -- very user-friendly but not
everyone will get it unless it spelled out for them)

eg should always be written "e.g."