Tweaks
Improved the caption for the image of the hidden drop-mode selector UI

Highlight important message about Autosave
Fixes typo about "recently accessed file".
Call it a "system tray" rather than just a tray.
Got rid of references to "the Editor" while retaining the message that
closing Grasshopper's editor does not result in quitting Grasshopper.
Mentioned the GrasshopperUnloadPlugin command (which can be helpful
during troubleshooting or when finished using Grasshopper for a
session).

Emphasized important warning about the Autosave feature (good to know
for those who would rely on it to build a huge file from scratch and
lose it all overnight due to a Windows update or something)

Improved caption about the drag-and-drop image.