Simplify and detail
to avoid calling the canvas a window (even though programmers think of
it as a 'window') since to most non-programmer users Grasshopper is in
one Window and Rhino is in another and together the two are comprised of
one window each, and each window as a bunch of stuff in it, but its just
one window, let's call it a portion of a window.
And to avoid the terminology "standard naming conventions" at least for
now (because I am not totally convinced we've accomplished that
everywhere yet (and which standard, exactly, anyway?)) just use the KIS
rule.
About the status bar, it does have more functionality than just
reporting the version of Grasshopper all day long.