How did that "and" get missed?
This file was good enough to leave alone except for the missing "and" in
line 21 but you know how when you reach down to pick up a speck of crumb
of the floor and look around while you are down there to see what else
you can pick up? The rest of the changes in this commit are kinda like
that.

Anyway, we don't want the users who don't have enough screen page to be
completely discouraged yet.  And we don't users on Vista or XP (without
Aerosnap installed) to already not trust our document because holding
the Windows key down and pressing left or right arrows doesn't work as
we said it would.