Didn't notice these were also clickable before
Of course you have to click while still holding down Ctrl or this dialog
will close.  This Ctrl+Tab screen has its limitations.  I did a quick
test and it doesn't work well if there are more open documents than the
number of thumbnails that will fit on the canvas. The drop-down list
invokes scroll bars making it the superior option.
I'm not going to mention any of this though.  After all, it's an
unlikely use-case.